requirements.txt
requests==2.25.0
mysql-connector-python==8.0.22